--- fourtooncookie/src/main/java/com/startingblue/fourtooncookie/aws/lambda/diaryimagegenerationpayload/DiaryImageGenerationCharacterPayload.java
@@ -1,4 +1,0 @@
-package com.startingblue.fourtooncookie.aws.lambda.diaryimagegenerationpayload;
-
-public record DiaryImageGenerationCharacterPayload(Long id, String name, String visionType, String basePrompt) {
-}
