--- fourtooncookie/src/main/java/com/startingblue/fourtooncookie/aws/lambda/diaryimagegenerationpayload/DiaryImageGenerationLambdaPayload.java
@@ -1,5 +1,0 @@
-package com.startingblue.fourtooncookie.aws.lambda.diaryimagegenerationpayload;
-
-public record DiaryImageGenerationLambdaPayload(Long diaryId, String content, DiaryImageGenerationCharacterPayload character) {
-
-}
