+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/diary/service/DiaryLambdaService.java
@@ -1,4 +1,4 @@
-package com.startingblue.fourtooncookie.aws.lambda.diaryimagegenerationpayload;
+package com.startingblue.fourtooncookie.diary.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -7,6 +7,7 @@ import com.startingblue.fourtooncookie.character.domain.Character;
 import com.startingblue.fourtooncookie.diary.domain.Diary;
 import com.startingblue.fourtooncookie.diary.domain.DiaryRepository;
 import com.startingblue.fourtooncookie.diary.domain.DiaryStatus;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
@@ -15,20 +16,17 @@ import software.amazon.awssdk.services.lambda.LambdaClient;
 import software.amazon.awssdk.services.lambda.model.InvocationType;
 
 @Service
+@RequiredArgsConstructor
 @Slf4j
-public class DiaryImageGenerationLambdaService extends LambdaService {
+public class DiaryLambdaService {
 
-    private static final String FUNCTION_NAME = "fourtooncookie-diaryimage-ai-apply-lambda";
+    private static final String IMAGE_GENERATION_FUNCTION_NAME = "fourtooncookie-diaryimage-ai-apply-lambda";
+    private static final InvocationType INVOCATION_TYPE = InvocationType.EVENT;
 
+    private final LambdaService lambdaService;
+    private final LambdaClient lambdaClient;
     private final ObjectMapper objectMapper;
-
     private final DiaryRepository diaryRepository;
-
-    public DiaryImageGenerationLambdaService(LambdaClient lambdaClient, ObjectMapper objectMapper, DiaryRepository diaryRepository) {
-        super(lambdaClient, FUNCTION_NAME, InvocationType.REQUEST_RESPONSE);
-        this.objectMapper = objectMapper;
-        this.diaryRepository = diaryRepository;
-    }
 
     @Async
     @Transactional
@@ -37,7 +35,7 @@ public class DiaryImageGenerationLambdaService extends LambdaService {
         try {
             DiaryImageGenerationLambdaPayload diaryImageGenerationLambdaPayload = buildPayload(diary, character);
             String serializePayload = serializePayload(diaryImageGenerationLambdaPayload);
-            invokeLambda(serializePayload);
+            lambdaService.invokeLambda(lambdaClient, IMAGE_GENERATION_FUNCTION_NAME, INVOCATION_TYPE, serializePayload);
             status = DiaryStatus.COMPLETED;
         } catch (Exception e) {
             log.error("Lambda 호출 중 오류 발생: {}", e.getMessage());
@@ -68,4 +66,9 @@ public class DiaryImageGenerationLambdaService extends LambdaService {
         diary.updateDiaryStatus(status);
         diaryRepository.save(diary);
     }
+
+    private record DiaryImageGenerationCharacterPayload(Long id, String name, String visionType, String basePrompt) { }
+
+    private record DiaryImageGenerationLambdaPayload(Long diaryId, String content, DiaryImageGenerationCharacterPayload character) { }
+
 }
